+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.78.0]
+
+### Added
+
+- default token "evaTknLogin" that will be replaced by current username
+
+### Fixed
+
+- fixed the management of custom colors of the pie component
+- fixed focusing on the reset button of the filter part
+- fixed the bug of dragging the graph component
+
 ## [1.77.0]
 
 ### Added
+
+- added field for choose interval tooltips for points multiline, created settings for change color lines and choose style line
 
 - added barplot display styles of multiline
 - updated sticky header and filter